pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "asdabot"
version = "0.1.0"
description = "CLI for autonomous grocery shopping on ASDA"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Mark Dunne" }]
keywords = ["asda", "groceries", "shopping", "delivery", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]
dependencies = [
    "algoliasearch>=4.37.1",
    "camoufox>=0.4.11",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "rich>=14.3.3",
    "tenacity>=9.1.4",
    "typer>=0.24.1",
]

[project.scripts]
asdabot = "asdabot.cli:main"

[project.urls]
Repository = "https://github.com/MarkDunne/AsdaBot"

[dependency-groups]
dev = [
    "ruff>=0.15.6",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
    "N",    # pep8-naming
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "TCH",  # flake8-type-checking
    "PIE",  # flake8-pie
    "PT",   # flake8-pytest-style
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "FLY",  # flynt
    "PERF", # perflint
    "FURB", # refurb
    "LOG",  # flake8-logging
    "PLC",  # pylint convention (catches non-top-level imports)
]
ignore = [
    "S101",  # assert is fine
    "T201",  # print is fine for CLI output
]

[tool.ruff.lint.per-file-ignores]
"asdabot/cli.py" = ["T201", "PLC0415"]  # lazy imports for checkout (avoids 1s startup)
